% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_simplify_bin_search.R
\name{bin_simplify}
\alias{bin_simplify}
\title{simplify curve to given nr data points}
\usage{
bin_simplify(path, size, precision)
}
\arguments{
\item{path}{This is a data.frame containing two columns with
information about the original trajectory/path. The columns are named x and y
and contain the x and y coordinates of the trajectory.}

\item{size}{This is the number of data points the simplified path should have}

\item{precision}{This is the error allowed on the number of data points. This
needs to be at least 1}
}
\description{
The function simplifies a polygonal curve P by an approximation
P'. Using a binary search on epsilon under the frechet distance, the desired
number of data points is returned.
}
