% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_discrete_frechet.R
\name{discrete_frechet}
\alias{discrete_frechet}
\title{compute discrete frechet distance}
\usage{
discrete_frechet(p1, p2)
}
\arguments{
\item{p1}{data.frame containing x and y coordinates of trajectory}

\item{p2}{data.frame containing x and y coordinates of trajectory}
}
\description{
from Thomas Eiter and Heikki Mannila
}
\references{
\url{http://www.kr.tuwien.ac.at/staff/eiter/et-archive/cdtr9464.pdf}
}
