% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_wrapper_functions.R
\name{dtw_dist}
\alias{dtw_dist}
\title{compute DTW distance between two paths}
\usage{
dtw_dist(p1, p2, sqr_dist, normalize_dist)
}
\arguments{
\item{p1}{This is a data.frame containing two columns with
information about the original trajectory/path. The columns are named x and y
and contain the x and y coordinates of the trajectory.}

\item{p2}{This is a data.frame containing two columns with
information about the original trajectory/path. The columns are named x and y
and contain the x and y coordinates of the trajectory.}

\item{sqr_dist}{default = FALSE. If dist_measure = 3, then:
square_dist = TRUE ensures the dtw uses squared distances
square_dist = FALSE ensures the dtw uses normal distances}

\item{normalize_dist}{default = TRUE. If dist_measure = 3, then:
norm_dist = TRUE ensures dtw distances returns a normalized distance
i.e. dtw_dist = dtw_dist/max(|p1|, |p2|)
if square_dist = TRUE then dtw_dist = sqrt(dtw_dist/max(|p1|, |p2|))
norm_dist = FALSE outputs actual dtw distance}
}
\description{
compute DTW distance between two paths
}
