% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_wrapper_functions.R
\name{frechet_distance}
\alias{frechet_distance}
\title{calculate continuous frechet distance}
\usage{
frechet_distance(p1, p2, precision)
}
\arguments{
\item{p1}{This is a data.frame containing two columns with
information about the original trajectory/path. The columns are named x and y
and contain the x and y coordinates of the trajectory.}

\item{p2}{This is a data.frame containing two columns with
information about the original trajectory/path. The columns are named x and y
and contain the x and y coordinates of the trajectory.}

\item{precision}{This is a number containing the precision with which to
calculate the frechet distance}
}
\value{
A number apx_eps which equals the frechet distance. Note that the
outcome is not exact. The true epsilon lies in the following range:
apx_eps - precision <= epsilon <= apx_eps
}
\description{
The function calculates the frechet distance between two paths.
The distance has a precision variable that indicates the precision of the
outcome.
}
