% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_fsp_centers.R
\name{fsp_centers}
\alias{fsp_centers}
\title{fsp centers}
\usage{
fsp_centers(paths, clust_out, dist_measure = 3, dtw_settings = c(1,
  TRUE, TRUE))
}
\arguments{
\item{paths}{This is a list of data.frames. Each data.frame represents a path.
Each data.frame contains two columns. The columns are named x and y
and contain the x and y coordinates of the trajectory}

\item{clust_out}{the output of one of the clustering algorithms}

\item{dist_measure}{input either "cont_frechet", "disc_frechet", "dtw or 1,2,3
respectively. Specifies the distance measure to be used.}

\item{dtw_settings}{a vector of 3 variables with settings when using dtw
distance c('method_dtw', 'squared_dist' ,' normalize_dist' )
method_dtw: integer 1,2,3,4; specified what center type to use
squared_dist: bool; specifies wheter to use squared distances
normalize_dist: bool; specifies wheteher to normalize distances to output in
cost in the same range as frechet distance}
}
\value{
Clustering output containing clusters, centers and result
}
\description{
fsp centers
}
