% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_kl_median.R
\name{kl_median}
\alias{kl_median}
\title{kl-median}
\usage{
kl_median(paths, clust_out, max_iter = 1, dist_measure = 2,
  clusts_to_use = NA, prec_eps = 1e-04, square_dist = FALSE,
  norm_dist = TRUE, update = TRUE)
}
\arguments{
\item{paths}{This is a list of data.frames. Each data.frame represents a path.
Each data.frame contains two columns. The columns are named x and y
and contain the x and y coordinates of the trajectory}

\item{clust_out}{the output of one of the clustering algorithms}

\item{max_iter}{integer with maximum number of iterations to perform}

\item{dist_measure}{either input 1, 2 or 3 correspondig to
"continuous frechet", "discrete frechet" or "dtw"}

\item{clusts_to_use}{(optional)vector of integers with indexes of clusters
to use in the algorithm#'}

\item{prec_eps}{This is a number containing the precision with which to
calculate the frechet distance}

\item{square_dist}{default = FALSE. If dist_measure = 3, then:
square_dist = TRUE ensures the dtw uses squared distances
square_dist = FALSE ensures the dtw uses normal distances}

\item{norm_dist}{default = TRUE. If dist_measure = 3, then:
norm_dist = TRUE ensures dtw distances returns a normalized distance
i.e. dtw_dist = dtw_dist/max(|p1|, |p2|)
if square_dist = TRUE then dtw_dist = sqrt(dtw_dist/max(|p1|, |p2|))
norm_dist = FALSE outputs actual dtw distance}

\item{update}{if TRUE then all output will be updated}
}
\value{
Clustering output containing clusters, centers and result
}
\description{
The function computes k centers of the given input paths and
initial centers, where the centers are simplified to l data points.
}
