% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clustering.R
\name{plot_all_clust}
\alias{plot_all_clust}
\title{plot all clusters without grid}
\usage{
plot_all_clust(all_paths, out_dt, name_plot = "")
}
\arguments{
\item{all_paths}{is the list of data.frames with the trajectories}

\item{out_dt}{contains the clustering output}

\item{name_plot}{is a string containing the title of the plot}
}
\description{
plot all clusters without grid
}
