% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clustering.R
\name{plot_all_clust_map}
\alias{plot_all_clust_map}
\title{plot all clusters on a mao using ggmap::ggmap and stamenmaps}
\usage{
plot_all_clust_map(all_paths, out_dt, name_plot = "", map_type,
  zoom_ggmap = 13)
}
\arguments{
\item{all_paths}{is the list of data.frames with the trajectories}

\item{out_dt}{contains the clustering output}

\item{name_plot}{is a string containing the title of the plot}

\item{map_type}{contains the map type you want to use, see
[stamenmaps](maps.stamen.com)}

\item{zoom_ggmap}{defines the quality of the map}
}
\description{
plot all clusters on a mao using ggmap::ggmap and stamenmaps
}
