% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_simplify_all.R
\name{simplify_all}
\alias{simplify_all}
\title{simplify curves to given nr data points}
\usage{
simplify_all(paths, size, precision)
}
\arguments{
\item{paths}{This is a list of data.frames. Each data.frame contains two
columns with information about the original trajectory/path. The columns are
named x and y and contain the x and y coordinates of the trajectory.}

\item{size}{This is the number of data points the simplified path should have}

\item{precision}{This is the error allowed on the number of data points. This
needs to be at least 1}
}
\description{
The function simplifies all polygonal curve P, stored in a list
by an approximation P'. Using a binary search on epsilon under the frechet
distance, the desired number of data points is returned for each path.
}
