% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_simplify_curve.R
\name{simplify_curve}
\alias{simplify_curve}
\title{simplify under frechet}
\usage{
simplify_curve(path, eps)
}
\arguments{
\item{path}{This is a data.frame containing two columns with
information about the original trajectory/path. The columns are named x and y
and contain the x and y coordinates of the trajectory.}

\item{eps}{This is a number representing epsilon which will be input for the
Frechet distance.}
}
\description{
The function simplifies a polygonal curve P by an approximation
P' subject to an error epsilon under the Frechet distance.
}
