% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_single_link.R
\name{single_link}
\alias{single_link}
\title{single linkage clustering}
\usage{
single_link(paths, radius = NA, k = NA, l = 20, prec_l = 1,
  dist_measure = 2, prec_eps = 1e-04, square_dist = FALSE,
  norm_dist = TRUE)
}
\arguments{
\item{paths}{This is a list of data.frames. Each data.frame represents a path.
Each data.frame contains two columns. The columns are named x and y
and contain the x and y coordinates of the trajectory}

\item{radius}{The algorithm computes the number of centers such that each
path's distance to one of the centers is smaller than or equal to radius}

\item{k}{The algorithm computs k centers}

\item{l}{The number of data points to simplify each center to.}

\item{prec_l}{This is a number containing the precision of the number of data
points to return.}

\item{dist_measure}{either input 1, 2 or 3 correspondig to
"continuous frechet", "discrete frechet" or "dtw"}

\item{prec_eps}{This is a number containing the precision with which to
calculate the frechet distance}

\item{square_dist}{default = FALSE. If dist_measure = 3, then:
square_dist = TRUE ensures the dtw uses squared distances
square_dist = FALSE ensures the dtw uses normal distances}

\item{norm_dist}{default = TRUE. If dist_measure = 3, then:
norm_dist = TRUE ensures dtw distances returns a normalized distance
i.e. dtw_dist = dtw_dist/max(|p1|, |p2|)
if square_dist = TRUE then dtw_dist = sqrt(dtw_dist/max(|p1|, |p2|))
norm_dist = FALSE outputs actual dtw distance}
}
\description{
This algorithm builds a hierarchy of clusterings and returns
a number of centers given either radius or k
}
