% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_move_transform.R
\name{transform_move}
\alias{transform_move}
\title{Transform movebank data to input format}
\usage{
transform_move(move_dt)
}
\arguments{
\item{move_dt}{move data structure, e.g. "Move", "MoveStack", "MoveBurst"}
}
\description{
this function transforms movebank data format to the right input
format for the clustering algorithms
}
